Added Enemy Ability Holder
Added a script named EnemyHolder that is attached to the enemy. The script handles when the enemy would use the ability. The logic is the same as AbilityHolder except the trigger is changed to when the player is detected. Currently, the enemy will repeatedly cast the ability after it detects the player (same logic as its movement).
Added a basic melee ability to the current enemy. The process is the same as adding ability to the player. Enemy's ability is using the same script  as the player does (MeleeAbility) but with different attributes (mainly cooldown and damage).
Modified MeleeAbility to make sure the ability is casted in the direction that the character is facing. May need small modification after sprites are set to right by default.

+++ Game/OtherSide/Assets/Scripts/Ability.cs
@@ -17,7 +17,7 @@ public class Ability : ScriptableObject
 
     [SerializeField] public float cooldownTime; // ability cooldown in seconds
     [SerializeField] public float activeTime; // ability duration in seconds
-    [SerializeField] public float damage; // ability damage to enemies
+    [SerializeField] public int damage; // ability damage to enemies
     [SerializeField] public int cost; // how much ability costs resource-wise, ints for now
 
     // ability range depends on the ability
